Refactoring Stage, Particles, Engine
In order to resolve circular dependency issues, Stage and Particles were changed to classes which are now part of the game object. Engine is not a class, but is also part of the game object now. It's a little messy right now but I forgot to actually test before the last commit and it was broken.